refactor: 532 - rename course-main to hero-course
