--- src/widgets/course-main/index.ts
@@ -1,3 +1,0 @@
-export type { CourseMainProps } from './ui/course-main';
-export { CourseMain } from './ui/course-main';
-export { loader } from './api/loader';
