+++ src/widgets/hero-course/api/loader.ts
