+++ src/widgets/hero-course/helpers/get-course-status.ts
