+++ src/widgets/hero-course/helpers/has-day.ts
