+++ src/widgets/hero-course/index.ts
@@ -1,0 +1,3 @@
+export type { CourseMainProps } from './ui/hero-course';
+export { CourseMain } from './ui/hero-course';
+export { loader } from './api/loader';
