+++ src/widgets/hero-course/ui/hero-course.module.scss
